requirements.txt
python-dotenv
bs4
lxml
selenium
discord